feat(settings): rename Branding to Portal and simplify welcome message
/admin/settings/portal is the real page; /branding redirects. Welcome
message is a single body field (legacy enabled+title folds into a heading
on read; disabled drafts become empty). Feedback nav is always on.
Settings nav: Help Center, Changelog, and Status are flat links.
