+++ apps/web/src/lib/server/domains/settings/__tests__/portal-welcome-card-config.test.ts
@@ -2,33 +2,21 @@ import { describe, it, expect } from 'vitest'
 import { parseJsonConfig } from '../settings.helpers'
 import {
   DEFAULT_PORTAL_CONFIG,
+  EMPTY_WELCOME_BODY,
   type PortalConfig,
   type PortalWelcomeCard,
   type PublicPortalConfig,
 } from '../settings.types'
 
 describe('PortalWelcomeCard defaults', () => {
-  it('is off by default', () => {
-    expect(DEFAULT_PORTAL_CONFIG.welcomeCard?.enabled).toBe(false)
-  })
-
-  it('has empty title by default', () => {
-    expect(DEFAULT_PORTAL_CONFIG.welcomeCard?.title).toBe('')
-  })
-
-  it('has an empty doc body by default', () => {
-    expect(DEFAULT_PORTAL_CONFIG.welcomeCard?.body).toEqual({
-      type: 'doc',
-      content: [{ type: 'paragraph' }],
-    })
+  it('has an empty doc body by default (card hidden)', () => {
+    expect(DEFAULT_PORTAL_CONFIG.welcomeCard?.body).toEqual(EMPTY_WELCOME_BODY)
   })
 })
 
 describe('PortalWelcomeCard type', () => {
   it('accepts a fully-specified welcome card', () => {
     const card: PortalWelcomeCard = {
-      enabled: true,
-      title: 'Share your product feedback!',
       body: {
         type: 'doc',
         content: [
@@ -39,25 +27,24 @@ describe('PortalWelcomeCard type', () => {
         ],
       },
     }
-    expect(card.enabled).toBe(true)
-    expect(card.title).toContain('Share')
+    expect(card.body.content?.[0]?.content?.[0]?.text).toContain('Tell us')
   })
 
   it('is exposed as an optional field on PortalConfig', () => {
     const cfg: PortalConfig = {
       ...DEFAULT_PORTAL_CONFIG,
-      welcomeCard: { enabled: true, title: 'Hi', body: { type: 'doc' } as never },
+      welcomeCard: { body: { type: 'doc' } as never },
     }
-    expect(cfg.welcomeCard?.enabled).toBe(true)
+    expect(cfg.welcomeCard?.body.type).toBe('doc')
   })
 
   it('is exposed on PublicPortalConfig so the portal SSR loader can read it', () => {
     const projection: PublicPortalConfig = {
       features: DEFAULT_PORTAL_CONFIG.features,
       openSignup: true,
-      welcomeCard: { enabled: true, title: 'x', body: { type: 'doc' } as never },
+      welcomeCard: { body: { type: 'doc' } as never },
     }
-    expect(projection.welcomeCard?.title).toBe('x')
+    expect(projection.welcomeCard?.body.type).toBe('doc')
   })
 })
 
@@ -68,13 +55,16 @@ describe('parseJsonConfig deep-merges welcomeCard', () => {
     expect(result.welcomeCard).toEqual(DEFAULT_PORTAL_CONFIG.welcomeCard)
   })
 
-  it('merges partial welcomeCard with defaults', () => {
+  it('merges a partial welcomeCard body over the empty default', () => {
     const stored = JSON.stringify({
-      welcomeCard: { enabled: true, title: 'Hello' },
+      welcomeCard: {
+        body: {
+          type: 'doc',
+          content: [{ type: 'paragraph', content: [{ type: 'text', text: 'Hello' }] }],
+        },
+      },
     })
     const result = parseJsonConfig(stored, DEFAULT_PORTAL_CONFIG)
-    expect(result.welcomeCard?.enabled).toBe(true)
-    expect(result.welcomeCard?.title).toBe('Hello')
-    expect(result.welcomeCard?.body).toEqual(DEFAULT_PORTAL_CONFIG.welcomeCard?.body)
+    expect(result.welcomeCard?.body.content?.[0]?.content?.[0]?.text).toBe('Hello')
   })
 })
