+++ apps/web/src/lib/server/domains/settings/__tests__/settings-cache.test.ts
@@ -218,8 +218,6 @@ describe('getWorkspaceSettings', () => {
       }),
       publicPortalConfig: {
         welcomeCard: {
-          enabled: true,
-          title: 'Hi',
           body: {
             type: 'doc',
             content: [{ type: 'image', attrs: { src: unsigned } }],
