+++ apps/web/src/lib/server/domains/settings/index.ts
@@ -42,7 +42,7 @@ export type {
 } from './settings.types'
 
 // Welcome card constants (no DB dependency)
-export { PORTAL_WELCOME_CARD_TITLE_MAX } from './settings.types'
+export { EMPTY_WELCOME_BODY } from './settings.types'
 
 // Default config values (no DB dependency)
 export {
