+++ apps/web/src/lib/server/domains/settings/settings.helpers.ts
@@ -4,14 +4,17 @@
  */
 import { db, eq, settings } from '@/lib/server/db'
 import { cacheDel, CACHE_KEYS } from '@/lib/server/cache'
-import { DomainException, InternalError, NotFoundError, ValidationError } from '@/lib/shared/errors'
+import { DomainException, InternalError, NotFoundError } from '@/lib/shared/errors'
 import { sanitizeTiptapContent } from '@/lib/server/sanitize-tiptap'
+import { isEmptyTiptapDoc } from '@/lib/shared/utils/is-empty-tiptap-doc'
 import { logger } from '@/lib/server/logger'
 import {
   DEFAULT_PORTAL_CONFIG,
-  PORTAL_WELCOME_CARD_TITLE_MAX,
+  EMPTY_WELCOME_BODY,
+  type PortalConfig,
   type PortalWelcomeCard,
 } from './settings.types'
+import type { TiptapContent } from '@/lib/shared/db-types'
 
 const log = logger.child({ component: 'settings-helpers' })
 
@@ -128,9 +131,70 @@ export async function writeMetadataKey(key: string, value: unknown): Promise<voi
 }
 
 /**
+ * Stored welcome card, including the legacy `{ enabled, title, body }`
+ * shape repaired by {@link resolveWelcomeCard} on read.
+ *
+ * @internal
+ */
+export type StoredWelcomeCard = {
+  enabled?: boolean
+  title?: string
+  body?: TiptapContent
+}
+
+/**
+ * Read-time repair of a stored welcome card to `{ body }`.
+ *
+ * - `enabled: true` and a non-empty title → prepend a level-2 heading
+ *   node to `body.content`, then drop `title` and `enabled`.
+ * - `enabled: false` → empty body (intentionally discards disabled drafts).
+ * - `enabled` absent (the post-simplification `{ body }` write) → body as stored.
+ *
+ * @internal
+ */
+export function resolveWelcomeCard(
+  card: StoredWelcomeCard | PortalWelcomeCard | undefined
+): PortalWelcomeCard {
+  if (!card) return { body: EMPTY_WELCOME_BODY }
+
+  const stored = card as StoredWelcomeCard
+  if (stored.enabled === false) return { body: EMPTY_WELCOME_BODY }
+
+  const body = stored.body ?? EMPTY_WELCOME_BODY
+  if (stored.enabled !== true) return { body }
+
+  const title = stored.title?.trim() ?? ''
+  if (!title) return { body }
+
+  const heading: TiptapContent = {
+    type: 'heading',
+    attrs: { level: 2 },
+    content: [{ type: 'text', text: title }],
+  }
+  return {
+    body: {
+      type: 'doc',
+      content: [heading, ...(body.content ?? [])],
+    },
+  }
+}
+
+/**
+ * Parse stored portalConfig and repair the welcome card to `{ body }`.
+ *
+ * @internal
+ */
+export function parsePortalConfig(json: string | null): PortalConfig {
+  const parsed = parseJsonConfig(json, DEFAULT_PORTAL_CONFIG)
+  return { ...parsed, welcomeCard: resolveWelcomeCard(parsed.welcomeCard) }
+}
+
+/**
  * Merge a partial `welcomeCard` update into the stored card. Unlike
  * {@link deepMerge}, the `body` field is replaced wholesale — a TipTap
  * doc with no `content` must clear the previous content, not retain it.
+ * The result is always the resolved `{ body }` shape (legacy enabled/title
+ * are dropped on write).
  *
  * @internal
  */
@@ -140,27 +204,27 @@ export function mergeWelcomeCard(
 ): PortalWelcomeCard {
   const base = existing ?? DEFAULT_PORTAL_CONFIG.welcomeCard!
   if (!input) return existing ?? base
-  return { ...base, ...input }
+  return { body: input.body ?? base.body }
 }
 
 /**
- * Project a stored welcome card for public consumption. Disabled cards
- * have draft title/body that must not leak through the public portal
- * config endpoint.
+ * Project a stored welcome card for public consumption. Empty bodies
+ * (including legacy disabled cards after {@link resolveWelcomeCard})
+ * are omitted so the portal renderer has nothing to show.
  *
  * @internal
  */
 export function publicWelcomeCard(
-  card: PortalWelcomeCard | undefined
+  card: StoredWelcomeCard | PortalWelcomeCard | undefined
 ): PortalWelcomeCard | undefined {
-  if (!card?.enabled) return undefined
-  return card
+  const resolved = resolveWelcomeCard(card)
+  if (isEmptyTiptapDoc(resolved.body)) return undefined
+  return resolved
 }
 
 /**
  * Normalize a partial `welcomeCard` update before it's merged into stored
- * portalConfig. Trims the title, enforces the length cap, and runs the
- * TipTap body through the standard sanitizer.
+ * portalConfig. Runs the TipTap body through the standard sanitizer.
  *
  * @internal
  */
@@ -169,16 +233,6 @@ export function normalizeWelcomeCardInput(
 ): Partial<PortalWelcomeCard> | undefined {
   if (!input) return input
   const normalized: Partial<PortalWelcomeCard> = { ...input }
-  if (typeof input.title === 'string') {
-    const trimmed = input.title.trim()
-    if (trimmed.length > PORTAL_WELCOME_CARD_TITLE_MAX) {
-      throw new ValidationError(
-        'WELCOME_CARD_TITLE_TOO_LONG',
-        `Welcome card title must be ${PORTAL_WELCOME_CARD_TITLE_MAX} characters or fewer`
-      )
-    }
-    normalized.title = trimmed
-  }
   if (input.body !== undefined) {
     normalized.body = sanitizeTiptapContent(input.body)
   }
