+++ apps/web/src/lib/shared/types/settings.ts
@@ -51,5 +51,5 @@ export {
   DEFAULT_AUTH_CONFIG,
   DEFAULT_PORTAL_CONFIG,
   DEFAULT_WIDGET_HOME_CARDS,
-  PORTAL_WELCOME_CARD_TITLE_MAX,
+  EMPTY_WELCOME_BODY,
 } from '@/lib/server/domains/settings/settings.types'
